deleting unused components
