--- app/801craftkitchen/page.js
@@ -1,3 +1,0 @@
-export default function Restaurant() {
-  return <h1>Craft Kitchen Page</h1>
-}
