--- app/brewery/ourbrews/page.js
@@ -1,9 +1,0 @@
-import { OurBrews } from "@/components/ourBrews";
-
-export default function Brews() {
-  return (
-    <div>
-      <OurBrews />
-    </div>
-  )
-}
