--- app/private-events/venue-options/page.js
@@ -1,9 +1,0 @@
-import { PrivateEventsVenueOptions } from "@/components/privateEventsVenueOptions";
-
-export default function VenueOptions() {
-  return (
-    <div>
-      <PrivateEventsVenueOptions />
-    </div>
-  )
-}
