--- components/UntappedIframe.jsx
@@ -1,7 +1,0 @@
-export function UntappedIframe() {
-  return (
-    <div>
-      <iframe src="https://untappd.com/v/beach-haus-brewery/2633342" frameBorder="0" width="100%" height="500px" allowFullScreen></iframe>
-    </div>
-  )
-}
