--- components/beerHallSemiPrivate.jsx
@@ -1,16 +1,0 @@
-import { EventsCarousel } from "./eventsCarousel";
-import reserved from "../public/BeerHallSemiPrivate/beerHallReserved.jpg";
-import weich from "../public/BeerHallSemiPrivate/DB16296C-80B7-42B0-9A18-0B0F84C65509.jpg";
-import xmas from "../public/BeerHallSemiPrivate/FullSizeRender.jpg";
-import jbs from "../public/BeerHallSemiPrivate/04BF669F-4982-475E-9E3C-B51B419FFD54.jpg";
-
-const IMAGES = [reserved, weich, xmas, jbs];
-const OPTIONS = { loop: true };
-
-export function BeerHallSemiPrivateCarousel() {
-  return (
-    <div>
-      <EventsCarousel images={IMAGES} options={OPTIONS} />
-    </div>
-  );
-}
