+++ components/privateEventsHeroCarousel.jsx
@@ -3,26 +3,6 @@
 import React, { useState, useEffect, useCallback } from "react";
 import useEmblaCarousel from "embla-carousel-react";
 import { Thumb } from "./emblaCarouselThumbsButton";
-import bhp1 from "../public/BeerHallPrivate/IMG_0788.jpeg";
-import bhp2 from "../public/BeerHallPrivate/IMG_1746.jpeg";
-import bhp3 from "../public/BeerHallPrivate/IMG_1747.jpeg";
-import bhp4 from "../public/BeerHallPrivate/IMG_3907.jpeg";
-import bhp5 from "../public/BeerHallPrivate/IMG_4132.jpeg";
-import bhp6 from "../public/BeerHallPrivate/IMG_6030.jpeg";
-import bhp7 from "../public/BeerHallPrivate/IMG_6537.jpg";
-import bhp8 from "../public/BeerHallPrivate/oT4e9E9c.jpeg";
-import r16_9_1 from "../public/16-9ratio/events-16:9.jpg";
-import r16_9_2 from "../public/16-9ratio/weich_16_9.jpg";
-import r16_9_27 from "../public/16-9ratio/62_27_img.jpeg";
-import r16_9_3 from "../public/16-9ratio/rooftopSunsetDrink16x9.jpeg";
-import r13_6_1 from "../public/13_6/bettertogeth.jpeg";
-import r13_6_2 from "../public/13_6/bluebeam.jpeg";
-import r13_6_3 from "../public/13_6/cheers.jpg";
-import r13_6_4 from "../public/13_6/cookiebeers.jpeg";
-import r13_6_5 from "../public/13_6/couple13_6.jpg";
-import r13_6_7 from "../public/13_6/donuts.jpg";
-import r13_6_8 from "../public/13_6/longDining.jpg";
-import r13_6_9 from "../public/13_6/thirty.jpeg";
 import slider_1 from "../public/13_6/decor.jpg";
 import slider_2 from "../public/16x9Crop/KLP02300.jpg";
 import slider_3 from "../public/16x9Crop/beerHall/oT4e9E9c.jpg";
