--- components/weddingsAtBeachHaus.jsx
@@ -1,10 +1,0 @@
-export function WeddingsAtBeachHaus() {
-  return (
-    <div className="pt-10">
-      <h1 class="text-center text-7xl">Weddings at Beach Haus</h1>
-      <p class="px-64 py-5 text-center text-sm">
-        Weddingwire Badge Code (script)
-      </p>
-    </div>
-  );
-}
